+++ guide/src/SUMMARY.md
@@ -3,6 +3,6 @@
 [Batch](./batch.md)
 - [Installation](./installation.md)
 - [Getting Started](./getting-started.md)
-- [Tasks](./tasks.md)
+- [Jobs](./jobs.md)
 - [Queries](./queries.md)
 - [Worker](./worker.md)
